+++ src/DesktopVirtualization/test/Get-AzWvdRegistrationInfo.Tests.ps1
@@ -30,7 +30,8 @@ Describe 'Get-AzWvdRegistrationInfo' {
                             -SsoContext $null `
                             -CustomRdpProperty $null `
                             -Ring $null `
-                            -ValidationEnvironment:$false
+                            -ValidationEnvironment:$false `
+                            -PreferredAppGroupType 'Desktop'
 
         $regInfo = Get-AzWvdRegistrationInfo -SubscriptionId $env.SubscriptionId `
                                     -ResourceGroupName $env.ResourceGroup `