+++ src/FrontDoor/FrontDoor/Cmdlets/NewFrontDoorBackendObject.cs
@@ -78,6 +78,18 @@ namespace Microsoft.Azure.Commands.FrontDoor.Cmdlets
         [Parameter(Mandatory = false, HelpMessage = "The Alias of the Private Link resource. Populating this optional field indicates that this backend is 'Private'")]
         public string PrivateLinkAlias { get; set; }
 
+        /// <summary>
+        /// The Resource ID of the Private Link. Populating this optional field indicates that this backend is 'Private'
+        /// </summary>
+        [Parameter(Mandatory = false, HelpMessage = "The Resource ID of the Private Link. Populating this optional field indicates that this backend is 'Private'")]
+        public string PrivateLinkResourceId { get; set; }
+
+        /// <summary>
+        /// The Location of Private Link resource. Location is required when PrivateLinkResourceId is set
+        /// </summary>
+        [Parameter(Mandatory = false, HelpMessage = "The Location of Private Link resource. Location is required when PrivateLinkResourceId is set")]
+        public string PrivateLinkLocation { get; set; }
+
         [Parameter(Mandatory = false, HelpMessage = "A custom message to be included in the approval request to connect to the Private Link")]
         public string PrivateLinkApprovalMessage { get; set; }
 
@@ -92,7 +104,9 @@ namespace Microsoft.Azure.Commands.FrontDoor.Cmdlets
                 Priority = !this.IsParameterBound(c => c.Priority) ? 1 : Priority,
                 Weight = !this.IsParameterBound(c => c.Weight) ? 50 : Weight,
                 BackendHostHeader = !this.IsParameterBound(c => c.BackendHostHeader) ? Address : BackendHostHeader,
-                PrivateLinkAlias = !this.IsParameterBound(c => c.PrivateLinkAlias) ? null : PrivateLinkAlias,
+                PrivateLinkAlias = PrivateLinkAlias,
+                PrivateLinkResourceId = PrivateLinkResourceId,
+                PrivateLinkLocation = PrivateLinkLocation,
                 PrivateLinkApprovalMessage = PrivateLinkApprovalMessage
             };
             WriteObject(Backend);