+++ src/HDInsight/HDInsight.Test/ScenarioTests/HDInsightClusterTests.cs
@@ -42,5 +42,19 @@ namespace Commands.HDInsight.Test.ScenarioTests
         {
             TestController.NewInstance.RunPowerShellTest(_logger, "Test-CmkClusterRelatedCommands");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestCreateClusterWithEncryptionInTransit()
+        {
+            TestController.NewInstance.RunPowerShellTest(_logger, "Test-CreateClusterWithEncryptionInTransit");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestCreateClusterWithPrivateLink()
+        {
+            TestController.NewInstance.RunPowerShellTest(_logger, "Test-CreateClusterWithPrivateLink");
+        }
     }
 }