+++ src/ApiManagement/ApiManagement/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 2.1.0
 * Added new `Add-AzApiManagementApiToGateway` cmdlet.
 * Added new `Get-AzApiManagementGateway` cmdlet.
 * Added new `Get-AzApiManagementGatewayHostnameConfiguration` cmdlet.